Replaced q-search with quasar-search
When trying to use `q-search`, I was getting errors in the console about the component not being registered. However, when I changed the component name to `quasar-search`, it worked as expected.

If this was a change on a specific version, let me know and I will update the docs to specify what to do depending on the version being used so that the docs are clear.