Add ES6 arrow function to initial Dialog example
The initial example, in the doc, of how to use a Dialog does not include an ES6 arrow in the handler.  This inevitably will result in people cutting and pasting the code and ending up with 'this' scope errors in the handler.   I changed the example to include the ES6 arrow by default, as well added a comment in the handler code itself so people 'cut and paste' that there is an explanation if needed.