Add a Example of advanced select object
I was thinking was not possible to add a sub-labels or other advanced things on q-select, due the lack of exemple on the documentation. Asking on the discord community other users didn't know also, so I would like to propose an update on documentation. Add a exemple with the "selectListOptions" array we have on the actual documentation.